Use the ActiveRecord connection pool for postgres
Using a raw PG connection loses any interaction or benefits offered by
ActiveRecord: no logging, no pool, no instrumentation, etc. Considering
that the library is only expected to be used in Rails applications this
drops the straight PG client in favor of a connection pool.
