+++ flippant.gemspec
@@ -20,6 +20,7 @@ Gem::Specification.new do |spec|
   spec.executables = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
 
+  spec.add_development_dependency "activerecord", "~> 5.1"
   spec.add_development_dependency "bundler"
   spec.add_development_dependency "pg", "~> 1.0"
   spec.add_development_dependency "rake"
